+++ src/main/java/gildedrose/shop/ItemResponse.java
@@ -1,4 +1,4 @@
 package gildedrose.shop;
 
-public record ItemResponse(String name, int sellIn, int quality, int value) {
+public record ItemResponse(String name, int sellIn, int quality, float value) {
 }
